// src/app/models/employee.model.ts
export interface Employee {
    emp_id:             number;
    emp_name:           string;
    emp_lastname:       string;
    emp_mail:           string;
    emp_datebirth:      string;
    emp_address:        string;
    emp_cellphone:      string;
    emp_vaccinestatus:  string;
    emp_vaccinename:    string;
    emp_vaccinedate:    string;
    emp_vaccinedose:    number;
    emp_user:           string;
    emp_password:       string;
    emp_identification: string;
}